correction de bug + arret clean avec ctrl-c

+++ game/2048.c
@@ -6,11 +6,19 @@ grid g;
 pid_t self_pid;
 pid_t display_pid;
 pid_t main_pid;
+
+pthread_t move_score;
+pthread_t goal;
+pthread_t main_thread;
+
 bool run = false;
 
 void end_game() {
     run = false;
     kill(main_pid, SIGRTMIN + 9);
+    pthread_kill(move_score, SIGRTMIN + 9);
+    pthread_kill(goal, SIGRTMIN + 9);
+    pthread_kill(main_thread, SIGRTMIN + 9);
 }
 
 void* thread_move_score(void* arg) {
@@ -20,8 +28,9 @@ void* thread_move_score(void* arg) {
     for (int i = 1; i <= 4; i++) sigaddset(&set, SIGRTMIN + i);
     sigaddset(&set, SIGRTMIN + 9);
 
+    int signum;
+
     while (run) {
-        int signum;
         sigwait(&set, &signum);
 
         if (signum != SIGRTMIN + 9) {
@@ -67,13 +76,15 @@ void* thread_goal(void* arg) {
 
         while (run) {
             sigwait(&set, &signum);
+
             state = check_game_state(&g);
-            
-            if (state != ONGOING) end_game();
 
             write(pipefd[1], &g, sizeof(grid));
+            if (signum == SIGRTMIN + 9) state = LOSE;
             write(pipefd[1], &state, sizeof(game_state));
             kill(self_pid, SIGRTMIN + 6);
+
+            if (state != ONGOING) end_game();
         }
         close(pipefd[1]);
         wait(NULL);
@@ -99,22 +110,18 @@ void* thread_goal(void* arg) {
         while (run) {
             sigwait(&set, &signum);
 
-            
             read(pipefd[0], &received_grid, sizeof(grid));
             read(pipefd[0], &state, sizeof(game_state));
 
+            system("clear");
+            printf("Score : %d\n\n", received_grid.score);
+            print_grid(received_grid.cells);
+            
+            if (state == WIN) printf("Félicitations ! Vous avez gagné !\n");
+            else if (state == LOSE) printf("Dommage ! Vous avez perdu !\n");
+            
             if (state != ONGOING) end_game(); // child and parent don't share run variable
-
-            if (run) {
-                system("clear");
-                printf("Score : %d\n\n", received_grid.score);
-                print_grid(received_grid.cells);
-
-                if (state == WIN) printf("Félicitations ! Vous avez gagné !\n");
-                else if (state == LOSE) printf("Dommage ! Vous avez perdu !\n");
-            }
-
-            kill(self_pid, SIGRTMIN + 8);
+            else kill(self_pid, SIGRTMIN + 8);
         }
         close(pipefd[0]);
         exit(EXIT_SUCCESS);
@@ -124,20 +131,18 @@ void* thread_goal(void* arg) {
 }
 
 void* thread_main(void* arg) {
-
-    pthread_t move_score;
-    pthread_t goal;
-
     pthread_create(&move_score, NULL, thread_move_score, NULL);
     pthread_create(&goal, NULL, thread_goal, NULL);
 
     int o = open("main_to_main", O_RDONLY);
     if (o == -1) {
+        end_game();
         perror("open main_to_main");
         exit(EXIT_FAILURE);
     }
     ssize_t bytes_read = read(o, &main_pid, sizeof(main_pid));
     if (bytes_read == -1) {
+        end_game();
         perror("read pid from main_to_main");
         exit(EXIT_FAILURE);
     }
@@ -146,6 +151,7 @@ void* thread_main(void* arg) {
     sigemptyset(&set);
     sigaddset(&set, SIGRTMIN + 6);
     sigaddset(&set, SIGRTMIN + 8);
+    sigaddset(&set, SIGRTMIN + 9);
     
     int signum;
     
@@ -154,33 +160,42 @@ void* thread_main(void* arg) {
     sigwait(&set, &signum);
 
     while (run) {
-        kill(main_pid, SIGRTMIN);
-        read(o, &direction, sizeof(direction));
-        switch (direction) {
-            case Up :
-                // Move Up
-                kill(self_pid, SIGRTMIN + 1);
-                break;
-            case Left :
-                // Move Left
-                kill(self_pid, SIGRTMIN + 2);
-                break;
-            case Down :
-                // Move Down
-                kill(self_pid, SIGRTMIN + 3);
-                break;
-            case Right :
-                // Move Right
-                kill(self_pid, SIGRTMIN + 4);
-                break;
-            default :
-                break;
+        if (signum != SIGRTMIN + 9) {
+            kill(main_pid, SIGRTMIN);
+            if (read(o, &direction, sizeof(direction)) == -1) {
+                end_game();
+                perror("read direction from main_to_main");
+                exit(EXIT_FAILURE);
+            }
+            switch (direction) {
+                case Up :
+                    // Move Up
+                    kill(self_pid, SIGRTMIN + 1);
+                    break;
+                case Left :
+                    // Move Left
+                    kill(self_pid, SIGRTMIN + 2);
+                    break;
+                case Down :
+                    // Move Down
+                    kill(self_pid, SIGRTMIN + 3);
+                    break;
+                case Right :
+                    // Move Right
+                    kill(self_pid, SIGRTMIN + 4);
+                    break;
+                default :
+                    end_game();
+                    break;
+            }
+            sigwait(&set, &signum);
+            if (signum != SIGRTMIN + 9) {
+                kill(display_pid, SIGRTMIN + 7);
+                sigwait(&set, &signum);
+            }
         }
-        sigwait(&set, &signum);
-        kill(display_pid, SIGRTMIN + 7);
-        sigwait(&set, &signum);
     }
-    pthread_kill(move_score, SIGRTMIN + 9);
+    
     pthread_join(move_score, NULL);
     pthread_join(goal, NULL);
 
@@ -203,9 +218,8 @@ int main() {
 
     self_pid = getpid();
 
-    pthread_t main;
-    pthread_create(&main, NULL, thread_main, NULL);
-    pthread_join(main, NULL);
+    pthread_create(&main_thread, NULL, thread_main, NULL);
+    pthread_join(main_thread, NULL);
 
     exit(EXIT_SUCCESS);
 }
